--- lib/ckb-utils/typed_data.dart
@@ -1,7 +1,0 @@
-import 'dart:typed_data';
-
-Uint8List uint8ListFromList(List<int> data) {
-  if (data is Uint8List) return data;
-
-  return Uint8List.fromList(data);
-}
