+++ lib/ckb_address.dart
@@ -1,9 +1,9 @@
 import 'dart:typed_data';
 
 import 'package:bip_bech32/bip_bech32.dart';
-import 'package:ckb_sdk/ckb-utils/crypto/crypto.dart';
-import 'package:ckb_sdk/ckb-utils/number.dart';
 import 'package:convert/convert.dart';
+
+import 'ckb_crypto.dart';
 
 part 'src/ckb_address/address_config.dart';
 part 'src/ckb_address/ckb_address.dart';
