+++ lib/ckb_system_contract.dart
@@ -1,7 +1,7 @@
 import 'package:ckb_sdk/ckb-rpc/ckb_api_client.dart';
-import 'package:ckb_sdk/ckb-utils/crypto/hash.dart';
 
 import 'ckb_address.dart';
+import 'ckb_crypto.dart';
 import 'ckb_types.dart';
 
 part 'src/ckb_system_contract/ckb_system_contract.dart';
