+++ lib/src/ckb_crypto/keccac.dart
@@ -1,8 +1,4 @@
-import 'dart:convert';
-import 'dart:typed_data';
-
-import 'package:ckb_sdk/ckb-utils/typed_data.dart';
-import 'package:pointycastle/digests/sha3.dart';
+part of 'package:ckb_sdk/ckb_crypto.dart';
 
 const int _shaBytes = 256 ~/ 8;
 // keccak is implemented as sha3 digest in pointycastle, see
