+++ lib/src/ckb_crypto/number.dart
@@ -1,17 +1,12 @@
-import 'dart:typed_data';
-
-import 'package:convert/convert.dart';
-import 'package:pointycastle/src/utils.dart' as pcUtils;
-
-import '../ckb_error.dart';
+part of 'package:ckb_sdk/ckb_crypto.dart';
 
 BigInt bytesToInt(List<int> bytes) => pcUtils.decodeBigInt(bytes);
 
 String remove0x(String hex) => hex.startsWith("0x") ? hex.substring(2) : hex;
 
-BigInt toBigInt(String hex) => BigInt.parse(hex, radix: 16);
+String hexAdd0x(String hex) => hex.startsWith("0x") ? hex : "0x$hex";
 
-String toHex(dynamic number, {bool pad = false, bool include0x = false, int forcePadLen}) {
+String numberToHex(dynamic number, {bool pad = false, bool include0x = false, int forcePadLen}) {
   String toHexSimple() {
     if (number is int)
       return number.toRadixString(16);
@@ -29,59 +24,13 @@ String toHex(dynamic number, {bool pad = false, bool include0x = false, int forc
   return hexString;
 }
 
-/// Converts the hexadecimal string, which can be prefixed with 0x, to a byte
-/// sequence.
-Uint8List hexToBytes(String hexStr) {
-  final bytes = hex.decode(remove0x(hexStr));
-  if (bytes is Uint8List) return bytes;
-
-  return Uint8List.fromList(bytes);
-}
-
 List<int> numberToBytes(dynamic number) =>
-    number is BigInt ? pcUtils.encodeBigInt(number) : hex.decode(toHex(number, pad: true));
+    number is BigInt ? pcUtils.encodeBigInt(number) : hex.decode(numberToHex(number, pad: true));
 
 String bytesToHex(List<int> bytes, {bool pad = false, bool include0x = false, int forcePadLen}) =>
-    toHex(bytesToInt(bytes), pad: pad, include0x: include0x, forcePadLen: forcePadLen);
-
-String hexAdd0x(String hex) => hex.startsWith("0x") ? hex : "0x$hex";
+    numberToHex(bytesToInt(bytes), pad: pad, include0x: include0x, forcePadLen: forcePadLen);
 
 List<int> intToBytes(BigInt number) => pcUtils.encodeBigInt(number);
-
-Uint8List hexStringToByteArray(String input) {
-  String cleanInput = remove0x(input);
-
-  int len = cleanInput.length;
-
-  if (len == 0) {
-    return Uint8List(0);
-  }
-  Uint8List data;
-  int startIdx;
-  if (len % 2 != 0) {
-    data = Uint8List((len ~/ 2) + 1);
-    data[0] = digitHex(cleanInput[0]);
-    startIdx = 1;
-  } else {
-    data = Uint8List((len ~/ 2));
-    startIdx = 0;
-  }
-
-  for (int i = startIdx; i < len; i += 2) {
-    data[(i + 1) ~/ 2] = (digitHex(cleanInput[i]) << 4) + digitHex(cleanInput[i + 1]);
-  }
-  return data;
-}
-
-int digitHex(String hex) {
-  int char = hex.codeUnitAt(0);
-  if (char >= '0'.codeUnitAt(0) && char <= '9'.codeUnitAt(0) ||
-      char >= 'a'.codeUnitAt(0) && char <= 'f'.codeUnitAt(0)) {
-    return int.parse(hex, radix: 16);
-  } else {
-    return -1;
-  }
-}
 
 String littleEndian(int number) {
   List<int> bytes = toBytesPadded(BigInt.from(number), 8);
@@ -124,4 +73,45 @@ List<int> arrayCopy(bytes, srcOffset, result, destOffset, bytesLength) {
     result[destOffset + i] = bytes[i];
   }
   return result;
+}
+
+Uint8List uint8ListFromList(List<int> data) {
+  if (data is Uint8List) return data;
+
+  return Uint8List.fromList(data);
+}
+
+Uint8List hexStringToByteArray(String input) {
+  String cleanInput = remove0x(input);
+
+  int len = cleanInput.length;
+
+  if (len == 0) {
+    return Uint8List(0);
+  }
+  Uint8List data;
+  int startIdx;
+  if (len % 2 != 0) {
+    data = Uint8List((len ~/ 2) + 1);
+    data[0] = digitHex(cleanInput[0]);
+    startIdx = 1;
+  } else {
+    data = Uint8List((len ~/ 2));
+    startIdx = 0;
+  }
+
+  for (int i = startIdx; i < len; i += 2) {
+    data[(i + 1) ~/ 2] = (digitHex(cleanInput[i]) << 4) + digitHex(cleanInput[i + 1]);
+  }
+  return data;
+}
+
+int digitHex(String hex) {
+  int char = hex.codeUnitAt(0);
+  if (char >= '0'.codeUnitAt(0) && char <= '9'.codeUnitAt(0) ||
+      char >= 'a'.codeUnitAt(0) && char <= 'f'.codeUnitAt(0)) {
+    return int.parse(hex, radix: 16);
+  } else {
+    return -1;
+  }
 }
