+++ test/address/address_test.dart
@@ -1,7 +1,6 @@
 import 'package:bip_bech32/bip_bech32.dart';
-import 'package:ckb_sdk/ckb-utils/crypto/crypto.dart';
-import 'package:ckb_sdk/ckb-utils/number.dart';
 import 'package:ckb_sdk/ckb_address.dart';
+import 'package:ckb_sdk/ckb_crypto.dart';
 import 'package:convert/convert.dart';
 import 'package:test/test.dart';
 
@@ -35,8 +34,8 @@ main() {
 
   test('privateKey to address TestNet', () {
     var ckbAddress = CKBAddress(Network.TestNet);
-    var address = ckbAddress.generate(bytesToHex(publicKeyFromPrivate(intToBytes(
-        toBigInt(remove0x("e79f3207ea4980b7fed79956d5934249ceac4751a4fae01a0f7c4a96884bc4e3"))))));
+    var address = ckbAddress.generate(bytesToHex(publicKeyFromPrivate(
+        hex.decode("e79f3207ea4980b7fed79956d5934249ceac4751a4fae01a0f7c4a96884bc4e3"))));
     expect(address, "ckt1q9gry5zgxmpjnmtrp4kww5r39frh2sm89tdt2l6v234ygf");
   });
