+++ test/type/lock_hash_test.dart
@@ -1,6 +1,6 @@
 import 'package:ckb_sdk/ckb-rpc/ckb_api_client.dart';
-import 'package:ckb_sdk/ckb-utils/crypto/crypto.dart';
 import 'package:ckb_sdk/ckb_address.dart';
+import 'package:ckb_sdk/ckb_crypto.dart';
 import 'package:ckb_sdk/ckb_system_contract.dart';
 import 'package:ckb_sdk/ckb_types.dart' show Script;
 import 'package:convert/convert.dart';
