+++ test/util/crypto_test.dart
@@ -1,9 +1,9 @@
-import 'package:ckb_sdk/ckb-utils/crypto/crypto.dart';
-import 'package:ckb_sdk/ckb-utils/number.dart';
-import 'package:test/test.dart';
 import 'dart:convert';
-import 'package:pointycastle/digests/blake2b.dart';
+
+import 'package:ckb_sdk/ckb_crypto.dart';
 import 'package:convert/convert.dart';
+import 'package:pointycastle/digests/blake2b.dart';
+import 'package:test/test.dart';
 
 main() {
   test('blake2b', () {
