+++ test/util/number_test.dart
@@ -1,4 +1,4 @@
-import 'package:ckb_sdk/ckb-utils/number.dart';
+import 'package:ckb_sdk/ckb_crypto.dart';
 import 'package:test/test.dart';
 
 main() {
